# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Nourishment 2.0 API",
        "contact": {
            "email": "marekt94@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/looseproductsinday": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of loose products in day by day ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "looseproductsinday"
                ],
                "summary": "List loose products in day",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Day ID",
                        "name": "dayId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/meal.LooseProductInDay"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update loose product in day by JSON body (must contain ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "looseproductsinday"
                ],
                "summary": "Update an existing loose product in day",
                "parameters": [
                    {
                        "description": "LooseProductInDay payload",
                        "name": "looseproduct",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.LooseProductInDay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new loose product in day from JSON body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "looseproductsinday"
                ],
                "summary": "Create a new loose product in day",
                "parameters": [
                    {
                        "description": "LooseProductInDay payload",
                        "name": "looseproduct",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.LooseProductInDay"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/looseproductsinday/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single loose product in day by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "looseproductsinday"
                ],
                "summary": "Get loose product in day by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LooseProductInDay ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meal.LooseProductInDay"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete loose product in day by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "looseproductsinday"
                ],
                "summary": "Delete a loose product in day",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "LooseProductInDay ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/meals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all meals",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "List meals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/meal.Meal"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update meal by JSON body (must contain ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Update an existing meal",
                "parameters": [
                    {
                        "description": "Meal payload",
                        "name": "meal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.Meal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new meal from JSON body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Create a new meal",
                "parameters": [
                    {
                        "description": "Meal payload",
                        "name": "meal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/meal.Meal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/meals/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single meal by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Get meal by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/meal.Meal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete meal by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meals"
                ],
                "summary": "Delete a meal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "meal.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "meal.LooseProductInDay": {
            "type": "object",
            "properties": {
                "dayId": {
                    "description": "ID_DNIA",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "product": {
                    "description": "ID_PRODUKTU jako relacja",
                    "allOf": [
                        {
                            "$ref": "#/definitions/meal.Product"
                        }
                    ]
                },
                "weight": {
                    "description": "ILOSC_W_G",
                    "type": "number"
                }
            }
        },
        "meal.Meal": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "productsInMeal": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meal.ProductInMeal"
                    }
                },
                "recipe": {
                    "type": "string"
                }
            }
        },
...
        "meal.Product": {
            "type": "object",
            "properties": {
                "carbohydrates": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/meal.Category"
                },
                "fat": {
                    "type": "number"
                },
                "fiber": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "kcalPer100": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "proteins": {
                    "type": "number"
                },
                "salt": {
                    "type": "number"
                },
                "sugar": {
                    "type": "number"
                },
                "sugarAndCarb": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "meal.ProductInMeal": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/meal.Product"
                },
                "weight": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Get token from authorization request and place in \"Value\" field \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}